Add functionality to get a proxy conn string, also husky and prettier.
To get the best proxy I need to order the list by the best
response times and then return the fastest. This means that
when we make calls to get the pages we will doing it with our
best possible option. Obviously due to the location of the proxy
sites hosting its response times will be biased I assume but
its the best approximation for now.

Also added husky and prettier for formatting and to run tests on commit
and push.