Move script into web folder and start refactoring
The idea is that this script becomes part of a web
api that allows a phone app to make requests to it.

To do this I need to do a few things:
- Change it so it accepts a story id parameter, not env var.
- Remove CLI specific code and seperate that off so it can
run as both a cli and an endpoint.
- Test it (which requires quite a bit of refactoring)

I have begun by splitting off the fictionpress scraping
code and testing it (not complete).