Use fictionpress scraper in cli, add calibre... not tested. Add timeout on requests of 10s.
The cli has been consistantly been having issues with
the proxy however the new scraper is fully tested so
just replaced all the old stuff with the new.

Add calibre folder as I intend to make a lib, not
finished or tested.

Add timeout to fictionpress requests because some of
the proxies take ages to time out and it just hangs.